requirements.txt
# Core dependencies
langgraph==0.2.50
langchain==0.3.13
langchain-core==0.3.28
langchain-community==0.3.13
langchain-openai==0.2.14
langchain-huggingface==0.1.2  # Kept for backward compatibility

# Data processing
polars==1.35.2
pyarrow==22.0.0
pandas==2.2.2

# HuggingFace
huggingface-hub>=1.1.4
transformers==4.48.1

# UI
streamlit==1.41.1

# Backend API
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.6

